feat: 모달 로직 훅 분리 및 리팩토링
